// app.client/src/api/models/CustomerDetailVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CustomerDetailVM = {
    customerId?: string | null;
    companyName?: string | null;
    description?: string | null;
    status?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    email?: string | null;
    phone?: string | null;
    paymentTerms?: string | null;
    taxId?: string | null;
    addressId?: string | null;
    isActive?: boolean;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
};