// app.client/src/api/models/AddressListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AddressListVM = {
    addressId?: string | null;
    name?: string | null;
    contactPersonName?: string | null;
    description?: string | null;
    email?: string | null;
    phone?: string | null;
    website?: string | null;
    addressLine1?: string | null;
    addressLine2?: string | null;
    city?: string | null;
    stateId?: string | null;
    countryId?: string | null;
    zipCode?: string | null;
    isActive?: boolean;
    category?: string | null;
    categoryId?: string | null;
    addressType?: string | null;
    addressTypeId?: string | null;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
};